back/requirement.txt
aenum==2.2.6
asgiref==3.3.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
dateparser==1.0.0
Django==3.1.5
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-cron==0.5.1
django-crontab==0.7.1
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
fbchat==1.9.7
idna==2.10
Markdown==3.3.3
oauthlib==3.1.0
paho-mqtt==1.5.1
pycparser==2.20
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tweepy==3.10.0
tzlocal==2.1
urllib3==1.26.3
